Cleaned up error messages by moving ones used in multiple modal files to one file in the discussion menus folder higher up
