+++ src/pieces/discussion/menu/DiscussionModalUtilities.ts
@@ -1,0 +1,3 @@
+export const INVALID_INPUT_PREFIX = "Invalid Input Format. Nothing Was Changed\n**Reasons(s):**";
+export const INPUT_ERROR_PREFIX = "Input Error. Nothing Was Changed\n**Reasons(s):**";
+export const SCORING_ERROR_MESSAGE = "Scoring Error; Contact Admin";
