+++ src/pieces/discussion/scoring/scoreFunctions.ts
@@ -5,8 +5,6 @@ import { userHasRoleWithId } from "../../../generalUtilities/RoleUtilities";
 import { getChannelInMainGuild } from "../../../generalUtilities/getChannelInMain";
 import { wait } from "../../../generalUtilities/wait";
 
-export const SCORING_ERROR_MESSAGE = "Scoring Error; Contact Admin";
-
 export async function scoreAllThreads(client: Client, forumId: string, discussionSpecs: DiscussionSpecs, staffId: string) {
 
     const threads = await getAllDiscussionThreads(client, forumId);
